requirements.txt
Flask==2.0.3
numpy
opencv_python_headless==4.5.5.64
tensorflow-cpu
tqdm==4.63.0
Werkzeug==2.0.3
gunicorn==20.1.0
opencv-contrib-python-headless
